requirements.txt
pandas
sqlalchemy
psycopg-binary
requests
python-dotenv
